milvus-demos-rag-retriever/requirements.txt
pymilvus
llama-index-embeddings-jinaai
llama-index-readers-file 
llama-index-vector-stores-milvus
llama-index-llms-together
python-dotenv
llama-index-llms-huggingface
llama-index-embeddings-huggingface
sentence-transformers
wget